lunarlander/requirements.txt
# Python dependencies for LunarLander PoC
# Tested with Python 3.10+

# Core RL libraries
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Machine learning
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0

# Optional: for GPU support
# torch>=2.0.0
